move tests to correct directory
