+++ prover/src/tables/ecdas.rs
@@ -23,7 +23,7 @@ use crate::tables::ecsm::ecdas_tuple;
 use ecsm::{EcdasStep, P_BYTES};
 
 /// `r = 3·p` as 33 little-endian bytes (the spec offset that keeps all quotients positive).
-const R_BYTES: [u8; 33] = [
+pub const R_BYTES: [u8; 33] = [
     0x8D, 0xF4, 0xFF, 0xFF, 0xFC, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
     0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
     0x02,
@@ -273,17 +273,17 @@ fn r_byte<F: IsField>(m: usize) -> FieldElement<F> {
 }
 
 #[derive(Clone, Copy)]
-enum Relation {
+pub enum Relation {
     Lambda,
     Xr,
     Yr,
 }
 
 /// Unconditional convolution carry constraint at limb `i`: `2^8·c_i − c_{i-1} − S_i = 0`.
-struct ConvCarry {
-    relation: Relation,
-    i: usize,
-    constraint_idx: usize,
+pub struct ConvCarry {
+    pub relation: Relation,
+    pub i: usize,
+    pub constraint_idx: usize,
 }
 
 impl ConvCarry {
@@ -391,9 +391,9 @@ impl TransitionConstraint<GoldilocksField, GoldilocksExtension> for ConvCarry {
 }
 
 /// `col = 0` (unconditional, degree 1). Used for the closing `c_63 = 0`.
-struct ColIsZero {
-    col: usize,
-    constraint_idx: usize,
+pub struct ColIsZero {
+    pub col: usize,
+    pub constraint_idx: usize,
 }
 
 impl TransitionConstraint<GoldilocksField, GoldilocksExtension> for ColIsZero {
@@ -413,11 +413,11 @@ impl TransitionConstraint<GoldilocksField, GoldilocksExtension> for ColIsZero {
 }
 
 /// `a · b = 0` or `a · (1 - b) = 0` (degree 2).
-struct MulZero {
-    a: usize,
-    b: usize,
-    b_complement: bool,
-    constraint_idx: usize,
+pub struct MulZero {
+    pub a: usize,
+    pub b: usize,
+    pub b_complement: bool,
+    pub constraint_idx: usize,
 }
 
 impl TransitionConstraint<GoldilocksField, GoldilocksExtension> for MulZero {
@@ -515,130 +515,4 @@ pub fn create_constraints(
     }
 
     (constraints, idx)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ecsm::compute_witness;
-
-    fn gx_le() -> [u8; 32] {
-        let mut be = [
-            0x79, 0xBE, 0x66, 0x7E, 0xF9, 0xDC, 0xBB, 0xAC, 0x55, 0xA0, 0x62, 0x95, 0xCE, 0x87,
-            0x0B, 0x07, 0x02, 0x9B, 0xFC, 0xDB, 0x2D, 0xCE, 0x28, 0xD9, 0x59, 0xF2, 0x81, 0x5B,
-            0x16, 0xF8, 0x17, 0x98,
-        ];
-        be.reverse();
-        be
-    }
-
-    fn k_le(v: u64) -> [u8; 32] {
-        let mut k = [0u8; 32];
-        k[..8].copy_from_slice(&v.to_le_bytes());
-        k
-    }
-
-    fn ops_for(k: u64) -> Vec<EcdasOperation> {
-        let w = compute_witness(&k_le(k), &gx_le()).unwrap();
-        w.steps
-            .into_iter()
-            .map(|step| EcdasOperation {
-                timestamp: 444,
-                step,
-            })
-            .collect()
-    }
-
-    fn row_view(
-        trace: &TraceTable<GoldilocksField, GoldilocksExtension>,
-        row: usize,
-    ) -> TableView<GoldilocksField, GoldilocksExtension> {
-        let main: Vec<FE> = (0..cols::NUM_COLUMNS)
-            .map(|c| *trace.main_table.get(row, c))
-            .collect();
-        TableView::new(vec![main], vec![])
-    }
-
-    #[test]
-    fn r_bytes_is_three_p() {
-        // 3·p as 33 little-endian bytes, cross-checked against the ecsm field modulus.
-        let p = ecsm::p();
-        let three_p = &p * 3u32;
-        let mut bytes = three_p.to_bytes_le();
-        bytes.resize(33, 0);
-        assert_eq!(&bytes[..], &R_BYTES[..]);
-    }
-
-    /// Every ECDAS constraint evaluates to zero on a generated trace across many scalars
-    /// (which exercise both double and add steps), including padding rows.
-    #[test]
-    fn constraints_hold_on_generated_trace() {
-        for k in [2u64, 3, 5, 7, 0xFF, 0xABCD, 1_000_003] {
-            let ops = ops_for(k);
-            assert!(!ops.is_empty(), "k={k} should have steps");
-            let trace = generate_ecdas_trace(&ops);
-
-            for row in 0..trace.num_rows() {
-                let view = row_view(&trace, row);
-                assert_eq!(
-                    IsBitConstraint::unconditional(cols::MU, 0).evaluate(&view),
-                    FE::zero(),
-                    "is_bit(mu) k={k} row {row}"
-                );
-                assert_eq!(
-                    IsBitConstraint::unconditional(cols::NEXT_OP, 0).evaluate(&view),
-                    FE::zero()
-                );
-                assert_eq!(
-                    MulZero {
-                        a: cols::OP,
-                        b: cols::NEXT_OP,
-                        b_complement: false,
-                        constraint_idx: 0
-                    }
-                    .evaluate(&view),
-                    FE::zero(),
-                    "op·next_op k={k} row {row}"
-                );
-                assert_eq!(
-                    MulZero {
-                        a: cols::NEXT_OP,
-                        b: cols::MU,
-                        b_complement: true,
-                        constraint_idx: 0
-                    }
-                    .evaluate(&view),
-                    FE::zero()
-                );
-                for relation in [Relation::Lambda, Relation::Xr, Relation::Yr] {
-                    for i in 0..64 {
-                        let v = ConvCarry {
-                            relation,
-                            i,
-                            constraint_idx: 0,
-                        }
-                        .evaluate(&view);
-                        assert_eq!(v, FE::zero(), "conv k={k} i={i} row {row}");
-                    }
-                }
-                for c_base in [cols::C0, cols::C1, cols::C2] {
-                    assert_eq!(
-                        ColIsZero {
-                            col: c_base + 63,
-                            constraint_idx: 0
-                        }
-                        .evaluate(&view),
-                        FE::zero()
-                    );
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn create_constraints_count() {
-        let (constraints, next) = create_constraints(0);
-        assert_eq!(constraints.len(), 199);
-        assert_eq!(next, 199);
-    }
 }
